+++ src/js/PlotBoilerplate.d.ts
@@ -54,7 +54,8 @@
  * @modified 2020-03-29 Fixed the enableSVGExport flag (read enableEport before).
  * @modified 2020-05-09 Included the Cirlcle class.
  * @modified 2020-06-22 Added the rasterScaleX and rasterScaleY config params.
- * @modified 2020-07-03 Fixed the selectedVerticesOnPolyon(Polygon) function: non-selectable vertices were selected too, before.
+ * @modified 2020-06-03 Fixed the selectedVerticesOnPolyon(Polygon) function: non-selectable vertices were selected too, before.
+ * @modified 2020-06-06 Replacing Touchy.js by AlloyFinger.js
  * @version  1.8.1
  *
  * @file PlotBoilerplate
@@ -585,6 +586,8 @@ export declare class PlotBoilerplate {
      * @return {void}
      **/
     private mouseWheelHandler;
+    private setOffset;
+    private setZoom;
     private installInputListeners;
     /**
      * Creates a control GUI (a dat.gui instance) for this