+++ src/js/PlotBoilerplate.js
@@ -55,7 +55,8 @@
  * @modified 2020-03-29 Fixed the enableSVGExport flag (read enableEport before).
  * @modified 2020-05-09 Included the Cirlcle class.
  * @modified 2020-06-22 Added the rasterScaleX and rasterScaleY config params.
- * @modified 2020-07-03 Fixed the selectedVerticesOnPolyon(Polygon) function: non-selectable vertices were selected too, before.
+ * @modified 2020-06-03 Fixed the selectedVerticesOnPolyon(Polygon) function: non-selectable vertices were selected too, before.
+ * @modified 2020-06-06 Replacing Touchy.js by AlloyFinger.js
  * @version  1.8.1
  *
  * @file PlotBoilerplate
@@ -441,7 +442,6 @@ var PlotBoilerplate = /** @class */ (function () {
         else if (drawable instanceof BezierPath_1.BezierPath) {
             this.drawables.push(drawable);
             var bezierPath = drawable;
-            // for( var i in bezierPath.bezierCurves ) {
             for (var i = 0; i < bezierPath.bezierCurves.length; i++) {
                 if (!drawable.adjustCircular && i == 0)
                     this.vertices.push(bezierPath.bezierCurves[i].startPoint);
@@ -457,13 +457,14 @@ var PlotBoilerplate = /** @class */ (function () {
             this.vertices.push(drawable.upperLeft);
             this.vertices.push(drawable.lowerRight);
             this.drawables.push(drawable);
+            // Todo: think about a IDragEvent interface
             drawable.upperLeft.listeners.addDragListener(function (e) {
                 drawable.lowerRight.add(e.params.dragAmount);
             });
             drawable.lowerRight.attr.selectable = false;
         }
         else {
-            throw "Cannot add drawable of unrecognized type."; // : " + drawable.constructor.name;
+            throw "Cannot add drawable of unrecognized type: " + (typeof drawable) + ".";
         }
         // This is a workaround for backwards compatibility when the 'redraw' param was not yet present.
         if (redraw || typeof redraw == 'undefined')
@@ -840,7 +841,6 @@ var PlotBoilerplate = /** @class */ (function () {
         var padding = parseFloat(window.getComputedStyle(container, null).getPropertyValue('padding')) || 0, border = parseFloat(window.getComputedStyle(canvas, null).getPropertyValue('border-width')) || 0, pl = parseFloat(window.getComputedStyle(container, null).getPropertyValue('padding-left')) || padding, pr = parseFloat(window.getComputedStyle(container, null).getPropertyValue('padding-right')) || padding, pt = parseFloat(window.getComputedStyle(container, null).getPropertyValue('padding-top')) || padding, pb = parseFloat(window.getComputedStyle(container, null).getPropertyValue('padding-bottom')) || padding, bl = parseFloat(window.getComputedStyle(canvas, null).getPropertyValue('border-left-width')) || border, br = parseFloat(window.getComputedStyle(canvas, null).getPropertyValue('border-right-width')) || border, bt = parseFloat(window.getComputedStyle(canvas, null).getPropertyValue('border-top-width')) || border, bb = parseFloat(window.getComputedStyle(canvas, null).getPropertyValue('border-bottom-width')) || border;
         var w = container.clientWidth;
         var h = container.clientHeight;
-        // console.log( 'w', w, 'h', h, 'border', border, 'padding', padding, pl, pr, pt, pb, bl, br, bt, bb );
         canvas.style.display = 'block';
         return { width: (w - pl - pr - bl - br), height: (h - pt - pb - bt - bb) };
     };
@@ -1064,10 +1064,11 @@ var PlotBoilerplate = /** @class */ (function () {
         //            until next keypress; the implication is, that [Ctrl] would still
         //            considered to be pressed which is not true.
         if (this.keyHandler.isDown('alt') || this.keyHandler.isDown('spacebar')) {
-            _self.draw.offset.add(e.params.dragAmount);
-            _self.fill.offset.set(_self.draw.offset);
+            /* _self.draw.offset.add( e.params.dragAmount );
+            _self.fill.offset.set( _self.draw.offset );
             _self.config.offsetX = _self.draw.offset.x;
-            _self.config.offsetY = _self.draw.offset.y;
+            _self.config.offsetY = _self.draw.offset.y; */
+            _self.setOffset(_self.draw.offset.clone().add(e.params.dragAmount));
             _self.redraw();
         }
         else {
@@ -1140,25 +1141,62 @@ var PlotBoilerplate = /** @class */ (function () {
         // CHANGED replaced _self by this
         var _self = this;
         var we = e;
-        var oldPos = _self.transformMousePosition(e.params.pos.x, e.params.pos.y);
+        // let oldPos : XYCoords = _self.transformMousePosition(e.params.pos.x, e.params.pos.y);
+        // let newPos : XYCoords = _self.transformMousePosition(e.params.pos.x, e.params.pos.y);
         if (we.deltaY < 0) {
-            _self.draw.scale.x = _self.fill.scale.x = _self.config.scaleX = _self.config.scaleX * zoomStep;
-            _self.draw.scale.y = _self.fill.scale.y = _self.config.scaleY = _self.config.scaleY * zoomStep;
+            //_self.draw.scale.x = _self.fill.scale.x = _self.config.scaleX = _self.config.scaleX*zoomStep;
+            //_self.draw.scale.y = _self.fill.scale.y = _self.config.scaleY = _self.config.scaleY*zoomStep;
+            _self.setZoom(_self.config.scaleX * zoomStep, _self.config.scaleY * zoomStep, new Vertex_1.Vertex(e.params.pos.x, e.params.pos.y)); // new Vertex(newOffsetX, newOffsetY) );
         }
         else if (we.deltaY > 0) {
-            _self.draw.scale.x = _self.fill.scale.x = _self.config.scaleX = Math.max(_self.config.scaleX / zoomStep, 0.01);
-            _self.draw.scale.y = _self.fill.scale.y = _self.config.scaleY = Math.max(_self.config.scaleY / zoomStep, 0.01);
+            //_self.draw.scale.x = _self.fill.scale.x = _self.config.scaleX = Math.max(_self.config.scaleX/zoomStep,0.01);
+            // _self.draw.scale.y = _self.fill.scale.y = _self.config.scaleY = Math.max(_self.config.scaleY/zoomStep,0.01);
+            _self.setZoom(_self.config.scaleX / zoomStep, _self.config.scaleY / zoomStep, new Vertex_1.Vertex(e.params.pos.x, e.params.pos.y)); // new Vertex(newOffsetx, newOffsetY) );
         }
-        var newPos = _self.transformMousePosition(e.params.pos.x, e.params.pos.y);
+        /* let newPos : XYCoords = _self.transformMousePosition(e.params.pos.x, e.params.pos.y);
         // Apply relative positioned zoom
-        var newOffsetX = _self.draw.offset.x + (newPos.x - oldPos.x) * _self.draw.scale.x;
-        var newOffsetY = _self.draw.offset.y + (newPos.y - oldPos.y) * _self.draw.scale.y;
+        let newOffsetX : number = _self.draw.offset.x + (newPos.x-oldPos.x)*_self.draw.scale.x;
+        let newOffsetY : number = _self.draw.offset.y + (newPos.y-oldPos.y)*_self.draw.scale.y;
         _self.draw.offset.x = _self.fill.offset.x = _self.config.offsetX = newOffsetX;
         _self.draw.offset.y = _self.fill.offset.y = _self.config.offsetY = newOffsetY;
+        */
         e.preventDefault();
         _self.redraw();
     };
     ;
+    /*
+    private getZoomX() : number {
+    this.draw.scale.x;
+    };
+
+    private getZoomY() : number {
+    this.draw.scale.y;
+    };
+    // TODO: test and doc
+    // combination with touch libraries?
+    private applyZoom(zoomFactorX, zoomFactorY) {
+    this.draw.scale.x = this.fill.scale.x = this.config.scaleX = Math.max(this.config.scaleX*zoomFactorX,0.01);
+    this.draw.scale.y = this.fill.scale.y = this.config.scaleY = Math.max(this.config.scaleY*zoomFactorY,0.01);
+    }
+    */
+    PlotBoilerplate.prototype.setOffset = function (newOffset) {
+        this.draw.offset.set(newOffset); // add( e.params.dragAmount );
+        this.fill.offset.set(newOffset); // _self.draw.offset );
+        this.config.offsetX = newOffset.x; // _self.draw.offset.x;
+        this.config.offsetY = newOffset.y; // _self.draw.offset.y;
+    };
+    ;
+    PlotBoilerplate.prototype.setZoom = function (zoomFactorX, zoomFactorY, interactionPos) {
+        // const oldPos = newPos;
+        var oldPos = this.transformMousePosition(interactionPos.x, interactionPos.y);
+        this.draw.scale.x = this.fill.scale.x = this.config.scaleX = Math.max(zoomFactorX, 0.01);
+        this.draw.scale.y = this.fill.scale.y = this.config.scaleY = Math.max(zoomFactorY, 0.01);
+        var newPos = this.transformMousePosition(interactionPos.x, interactionPos.y);
+        var newOffsetX = this.draw.offset.x + (newPos.x - oldPos.x) * this.draw.scale.x;
+        var newOffsetY = this.draw.offset.y + (newPos.y - oldPos.y) * this.draw.scale.y;
+        this.draw.offset.x = this.fill.offset.x = this.config.offsetX = newOffsetX;
+        this.draw.offset.y = this.fill.offset.y = this.config.offsetY = newOffsetY;
+    };
     PlotBoilerplate.prototype.installInputListeners = function () {
         var _self = this;
         if (this.config.enableMouse) {
@@ -1181,22 +1219,17 @@ var PlotBoilerplate = /** @class */ (function () {
         }
         if (this.config.enableTouch) {
             // Install a touch handler on the canvas.
-            if (!window["Touchy"] || typeof window["Touchy"] != "function") {
-                console.warn("Cannot initialize the touch handler. Touchy is missig. Did you include it?");
-            }
-            else {
-                // Convert absolute touch positions to relative DOM element position (relative to canvas)
-                var relPos_1 = function (pos) {
-                    return { x: pos.x - _self.canvas.offsetLeft,
-                        y: pos.y - _self.canvas.offsetTop
-                    };
+            var relPos_1 = function (pos) {
+                return { x: pos.x - _self.canvas.offsetLeft,
+                    y: pos.y - _self.canvas.offsetTop
                 };
+            };
+            if (window["Touchy"] && typeof window["Touchy"] == "function") {
+                // Convert absolute touch positions to relative DOM element position (relative to canvas)
                 // Some private vars to store the current mouse/position/button state.
                 var touchMovePos = null;
                 var touchDownPos = null;
                 var draggedElement = null;
-                // TODO
-                // ERROR, THIS DOES NOT COMPILE PROPERLY WITH TYPESCRIPT.
                 var Touchy = (window["Touchy"]);
                 new Touchy(this.canvas, { one: function (hand, finger) {
                         touchMovePos = new Vertex_1.Vertex(relPos_1(finger.lastPoint));
@@ -1223,14 +1256,124 @@ var PlotBoilerplate = /** @class */ (function () {
                         }
                     }
                 });
-            } // END else
+            }
+            else if (window["AlloyFinger"] && typeof window["AlloyFinger"] == "function") {
+                console.log('Alloy finger found.');
+                // Do not include AlloyFinger itself to the library
+                // (17kb, but we want to keep this lib as tiny as possible).
+                var AF = window["AlloyFinger"];
+                var touchMovePos = null;
+                var touchDownPos = null;
+                var draggedElement = null;
+                var multiTouchStartScale = null;
+                var af = new AF(this.canvas, {
+                    touchStart: function (e) {
+                        //console.log('touchStart',e);
+                        if (e.touches.length == 1) {
+                            touchMovePos = new Vertex_1.Vertex(relPos_1({ x: e.touches[0].clientX, y: e.touches[0].clientY }));
+                            touchDownPos = new Vertex_1.Vertex(relPos_1({ x: e.touches[0].clientX, y: e.touches[0].clientY }));
+                            draggedElement = _self.locatePointNear(_self.transformMousePosition(touchMovePos.x, touchMovePos.y), PlotBoilerplate.DEFAULT_TOUCH_TOLERANCE / Math.min(_self.config.cssScaleX, _self.config.cssScaleY));
+                        }
+                    },
+                    touchMove: function (e) {
+                        // console.log('touchMove',e);
+                        // updateStatus( "#touches: " + e.touches.length + ", dX=" + e.deltaX + ", dY=" + e.deltaY );
+                        if (e.touches.length == 1 && draggedElement) {
+                            e.preventDefault();
+                            e.stopPropagation();
+                            // The Touchy-points also have 'id' and 'time' attributes
+                            // which we are not interested in here.
+                            // hand.on('move', (points:Array<XYCoords>) => { 
+                            var rel = relPos_1({ x: e.touches[0].clientX, y: e.touches[0].clientY }); //  points[0] );
+                            var trans = _self.transformMousePosition(rel.x, rel.y);
+                            var diff = new Vertex_1.Vertex(_self.transformMousePosition(touchMovePos.x, touchMovePos.y)).difference(trans);
+                            if (draggedElement.typeName == 'vertex') {
+                                if (!_self.vertices[draggedElement.vindex].attr.draggable)
+                                    return;
+                                _self.vertices[draggedElement.vindex].add(diff);
+                                var draggingVertex = _self.vertices[draggedElement.vindex];
+                                var fakeEvent = { params: { dragAmount: diff.clone(), wasDragged: true, mouseDownPos: touchDownPos.clone(), mouseDragPos: touchDownPos.clone().add(diff), vertex: draggingVertex } };
+                                draggingVertex.listeners.fireDragEvent(fakeEvent);
+                                _self.redraw();
+                            }
+                            touchMovePos = new Vertex_1.Vertex(rel);
+                            // } );
+                        }
+                        else if (e.touches.length == 2) {
+                            // If at least two fingers touch and move, then change the draw offset (panning).
+                            e.preventDefault();
+                            e.stopPropagation();
+                            _self.setOffset(_self.draw.offset.clone().addXY(e.deltaX, e.deltaY)); // Apply zoom?
+                            _self.redraw();
+                        }
+                    },
+                    touchEnd: function (e) {
+                        //console.log('touchEnd',e);
+                        touchMovePos = null;
+                        touchDownPos = null;
+                        draggedElement = null;
+                        multiTouchStartScale = null;
+                    },
+                    touchCancel: function (e) {
+                        //console.log('touchCancel',e);
+                        touchMovePos = null;
+                        touchDownPos = null;
+                        draggedElement = null;
+                        multiTouchStartScale = null;
+                    },
+                    multipointStart: function (e) {
+                        console.log('multipointStart', e);
+                        multiTouchStartScale = _self.draw.scale.clone();
+                        // updateStatus( "" );
+                    },
+                    multipointEnd: function (e) {
+                        //console.log('multipointEnd',e);
+                        multiTouchStartScale = null;
+                    },
+                    /* tap: function (e) {
+                    //console.log('tap',e);
+                    },
+                    doubleTap: function (e) {
+                    //console.log('doubleTap',e);
+                    },
+                    longTap: function (e) {
+                    //console.log('longTap',e);
+                    },
+                    singleTap: function (e) {
+                    //console.log('singleTap',e);
+                    },
+                    rotate: function (e) {
+                    //console.log(e.angle);
+                    }, */
+                    pinch: function (e) {
+                        var fingerA = new Vertex_1.Vertex(e.touches.item(0).clientX, e.touches.item(0).clientY);
+                        var fingerB = new Vertex_1.Vertex(e.touches.item(1).clientX, e.touches.item(1).clientY);
+                        var center = new Line_1.Line(fingerA, fingerB).vertAt(0.5);
+                        // updateStatus("Zoom: " + e.zoom + ", " + fingerA + ", " + fingerB);
+                        // console.log(e.zoom);
+                        // window.alert(e.zoom);
+                        // this.zoom(e.zoom);
+                        _self.setZoom(multiTouchStartScale.x * e.zoom, multiTouchStartScale.y * e.zoom, center);
+                        _self.redraw();
+                    }
+                    /* pressMove: function (e) {
+                    //console.log(e.deltaX);
+                    //console.log(e.deltaY);
+                    },
+                    swipe: function (e) {
+                    //console.log("swipe" + e.direction);
+                    } */
+                });
+            }
+            else {
+                console.warn("Cannot initialize the touch handler. Touchy and AlloyFinger are missig. Did you include at least one of them?");
+            } // END 
         }
         else {
             _self.console.log('Touch interaction disabled.');
         }
         if (this.config.enableKeys) {
             // Install key handler
-            // var keyHandler : KeyHandler = new KeyHandler( { trackAll : true } )
             this.keyHandler = new KeyHandler_1.KeyHandler({ trackAll: true })
                 .down('escape', function () {
                 _self.clearSelection(true);